[refactor]: data 구조 수정에 따른 컴포넌트 수정 (#33)
